+++ src/main/java/com/odde/massivemailer/service/TrackingService.java
@@ -1,5 +1,5 @@
 package com.odde.massivemailer.service;
 
 public interface TrackingService {
-    int updateViewCount(Long token);
+    int updateViewCount(Integer token);
 }
